+++ modules/tools/include/ifm3d/tools/ovp8xx/get_service_report_app.h
@@ -25,6 +25,8 @@ namespace ifm3d
     virtual void Execute(CLI::App* app) override;
     virtual CLI::App* CreateCommand(CLI::App* parent) override;
 
+    std::string outfile = "service_report.zip";
+
   }; // end: class GetServiceReportApp
 } // end: namespace ifm3d